change directory & files to mimic the package
