+++ app/remix-pwa-sw/index.ts
@@ -1,0 +1,11 @@
+export {
+    matchRequest
+} from './fetch'
+
+export {
+    isAssetRequest,
+    isDocumentRequest,
+    isLoaderRequest
+} from './core'
+
+export * from './strategy'
