+++ app/remix-pwa-sw/react/index.ts
@@ -1,0 +1,2 @@
+export { useSWEffect } from './sw-hook'
+export { loadServiceWorker } from './loader'
