+++ app/remix-pwa-sw/react/loader.ts
