+++ app/remix-pwa-sw/react/sw-hook.ts
