+++ app/remix-pwa-sw/strategy.ts
